reorganize files for component context
