+++ projects/ng-dev/src/lib/test-context/component-context/component-context-next.spec.ts
@@ -1,15 +1,12 @@
 import {
   Component,
-  ElementRef,
   InjectionToken,
   Input,
   NgModule,
   OnChanges,
   SimpleChanges,
-  ViewChild,
 } from '@angular/core';
 import { ComponentFixture } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
 import {
   ANIMATION_MODULE_TYPE,
   BrowserAnimationsModule,
@@ -78,75 +75,6 @@ describe('ComponentContextNext', () => {
         expect(ctx.inject(ANIMATION_MODULE_TYPE)).toBe('NoopAnimations');
       });
     });
-
-    it('uses the components selector if it is a tag name', () => {
-      @Component({ selector: 's-tag-name', template: '' })
-      class TagNameComponent {}
-
-      const ctx = new ComponentContextNext(TagNameComponent);
-      ctx.run(() => {
-        expect(
-          ctx.fixture.debugElement.query(By.directive(TagNameComponent)).name,
-        ).toBe('s-tag-name');
-      });
-    });
-
-    it("can handle components that don't have a selector", () => {
-      @Component({ template: 'the template' })
-      class NoSelectorComponent {}
-
-      const ctx = new ComponentContextNext(NoSelectorComponent);
-      ctx.run(() => {
-        expect(ctx.fixture.nativeElement.textContent).toContain('the template');
-      });
-    });
-
-    it('can handle components whose selectors are not tag names', () => {
-      // tslint:disable-next-line:component-selector
-      @Component({ selector: '[myAttribute]', template: 'the template' })
-      class AttributeSelectorComponent {}
-      const ctx = new ComponentContextNext(AttributeSelectorComponent);
-      ctx.run(() => {
-        expect(ctx.fixture.nativeElement.textContent).toContain('the template');
-      });
-    });
-
-    it('picks up inputs that are setters', () => {
-      @Component({ template: '' })
-      class SetterInputComponent {
-        receivedValue?: string;
-
-        @Input() set setterInput(value: string) {
-          this.receivedValue = value;
-        }
-      }
-
-      const ctx = new ComponentContextNext(SetterInputComponent);
-      ctx.run({ inputs: { setterInput: 'sent value' } }, () => {
-        expect(ctx.getComponentInstance().receivedValue).toBe('sent value');
-      });
-    });
-
-    it("can handle components that don't have inputs", () => {
-      @Component({ template: '' })
-      class NoInputComponent {}
-
-      expect(() => {
-        new ComponentContextNext(NoInputComponent).run(noop);
-      }).not.toThrowError();
-    });
-
-    it('can handle components that use ViewChild in tricky ways', () => {
-      @Component({ template: '' })
-      class TrickyViewChildComponent {
-        @Input() tricky?: string;
-        @ViewChild('tricky') trickyChild!: ElementRef;
-      }
-      const ctx = new ComponentContextNext(TrickyViewChildComponent);
-      ctx.run({ inputs: { tricky: 'the value' } }, () => {
-        expect(ctx.getComponentInstance().tricky).toBe('the value');
-      });
-    });
   });
 
   describe('.init()', () => {
@@ -182,19 +110,6 @@ describe('ComponentContextNext', () => {
         expect(ngOnChangesSpy).toHaveBeenCalledTimes(1);
       });
     });
-
-    it('can update renamed inputs', () => {
-      @Component({ template: '{{ propertyName }}' })
-      class RenamedInputComponent {
-        // tslint:disable-next-line:no-input-rename
-        @Input('bindingName') propertyName?: string;
-      }
-
-      const ctx = new ComponentContextNext(RenamedInputComponent);
-      ctx.run({ inputs: { propertyName: 'custom value' } }, () => {
-        expect(ctx.fixture.nativeElement.textContent).toContain('custom value');
-      });
-    });
   });
 
   describe('.getComponentInstance()', () => {
@@ -224,21 +139,6 @@ describe('ComponentContextNext', () => {
         const changes: SimpleChanges = ngOnChangesSpy.calls.mostRecent()
           .args[0];
         expect(changes.myInput.currentValue).toBe('new value');
-      });
-    });
-
-    it('throws an error if given a non-input', () => {
-      @Component({ template: '' })
-      class NonInputComponent {
-        // tslint:disable-next-line:no-input-rename
-        @Input('nonInputBindingName') nonInput?: string;
-      }
-
-      const ctx = new ComponentContextNext(NonInputComponent);
-      ctx.run(() => {
-        expect(() => {
-          ctx.updateInputs({ nonInput: 'value' });
-        }).toThrowError('"nonInput" is not an input for this component');
       });
     });
   });
