+++ projects/ng-dev/src/lib/test-context/component-context/component-context-next.ts
@@ -1,4 +1,4 @@
-import { Component, Type } from '@angular/core';
+import { Type } from '@angular/core';
 import {
   ComponentFixture,
   TestBed,
@@ -6,10 +6,10 @@ import {
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { assert } from '@s-libs/js-core';
-import { flatten, keys, map } from '@s-libs/micro-dash';
-import { trimLeftoverStyles } from '../trim-leftover-styles';
-import { AngularContext, extendMetadata } from './angular-context';
+import { keys } from '@s-libs/micro-dash';
+import { trimLeftoverStyles } from '../../trim-leftover-styles';
+import { AngularContext, extendMetadata } from '../angular-context';
+import { createDynamicWrapper, getInputs } from './dynamic-wrapper.component';
 
 /** @hidden */
 export interface ComponentContextNextInit<ComponentType> {
@@ -121,68 +121,4 @@ export class ComponentContextNext<
     this.fixture.destroy();
     super.cleanUp();
   }
-}
-
-function createDynamicWrapper<T>(componentType: Type<T>): Type<T> {
-  const selector = getSelector(componentType);
-  const inputs = getInputs(componentType);
-
-  const template = `
-    <div>
-      <${selector}
-        ${inputs
-          .map(({ binding, property }) => `[${binding}]="${property}"`)
-          .join(' ')}
-      ></${selector}>
-    </div>
-  `;
-
-  @Component({ template })
-  class DynamicWrapperComponent {}
-
-  return DynamicWrapperComponent as Type<T>;
-}
-
-function getSelector(componentType: Type<unknown>): string {
-  const annotations = Reflect.getOwnPropertyDescriptor(
-    componentType,
-    '__annotations__',
-  );
-  assert(annotations, 'That does not appear to be a component');
-
-  let selector = annotations.value.find((decorator: any) => decorator.selector)
-    ?.selector;
-  if (!isValidSelector(selector)) {
-    selector = 's-libs-component-under-test';
-    TestBed.overrideComponent(componentType, { set: { selector } });
-  }
-  return selector;
-}
-
-function isValidSelector(selector: string): boolean {
-  if (!selector) {
-    return false;
-  }
-  try {
-    document.createElement(selector);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-function getInputs(
-  componentType: any,
-): Array<{ binding: string; property: string }> {
-  return flatten(
-    map(componentType.propDecorators, (decorators: any[], property: string) => {
-      const inputDecorators = decorators.filter(
-        (decorator) => decorator.type.prototype.ngMetadataName === 'Input',
-      );
-      return inputDecorators.map((decorator) => {
-        const binding = decorator.args?.[0] || property;
-        return { property, binding };
-      });
-    }),
-  );
 }
