+++ projects/ng-dev/src/lib/test-context/component-context/index.ts
@@ -1,0 +1,5 @@
+export { ComponentContext, ComponentContextInit } from './component-context';
+export {
+  ComponentContextNext,
+  ComponentContextNextInit,
+} from './component-context-next';
