+++ projects/ng-dev/src/lib/test-context/index.ts
@@ -1,7 +1,3 @@
+export * from './component-context';
 export { AngularContext } from './angular-context';
-export { ComponentContext, ComponentContextInit } from './component-context';
-export {
-  ComponentContextNext,
-  ComponentContextNextInit,
-} from './component-context-next';
 export { Synchronized } from './synchronize';
